requirements.txt
absl-py==0.15.0
adabelief-pytorch==0.2.1
adabelief-tf==0.2.1
aiofile==3.7.4
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
albumentations==0.4.6
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
cachetools==4.2.4
caio==0.9.3
calamari-ocr==2.1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coremltools==5.1.0
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.5.5
doxapy @ git+https://gitlab2.informatik.uni-wuerzburg.de/nof36hn/binarizer-doxa.git@175045c982494e1fc50c7a6046d5dde091d3c3cf
edit-distance==1.0.4
editdistance==0.6.0
efficientnet-pytorch==0.6.3
et-xmlfile==1.1.0
ezodf==0.3.2
flake8==4.0.1
flatbuffers==1.12
fonttools==4.28.5
frozenlist==1.2.0
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.2.0
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kraken==3.0.6
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib==3.5.1
mccabe==0.6.1
mpmath==1.2.1
multidict==5.2.0
munch==2.5.0
mypy-extensions==0.4.3
networkx==2.6.3
nptyping==1.4.4
numpy==1.19.2
oauthlib==3.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
PageXML-to-Mask-converter==0.1
paiargparse==1.1.2
pandas==1.3.5
Pillow==6.2.2
pooch==1.4.0
pretrainedmodels==0.7.4
prettytable==2.5.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.6.3
seaborn==0.9.1
segmentation-models-pytorch==0.2.1
Shapely==1.8.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
sympy==1.9
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tfaip==1.2.5
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
ttach==0.0.3
typeguard==2.13.3
typing-extensions==3.7.4.3
typing-inspect==0.7.1
typish==1.9.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==3.0.2
yarl==1.7.2
zipp==3.6.0
